requirements.txt
MetaTrader5
pandas
numpy
matplotlib
scikit-learn
joblib
python-dotenv
pytest